refactor(fonts): centralize font configuration and clean up font-related styles
Consolidate font configuration into a single module and remove redundant font-family declarations across components. This improves maintainability and ensures consistent font usage throughout the application.
