+++ src/ui/components/InputWithLabel/input.module.scss
@@ -1,25 +1,15 @@
 .root {
   position: relative;
-  /* font-family: "Switzer" !important; */
-  font-family: var(--mantine-font-family) !important;
 }
 
 .input {
   height: 54px;
   padding-top: 25px;
-  // padding-top: 18px;
-  // font-family: "Switzer" !important;
-  font-family: var(--mantine-font-family) !important;
   background-color: #fcfcfd;
   border: 1px solid #f2f4f7;
-
-  &::placeholder {
-    font-family: var(--mantine-font-family) !important;
-  }
 }
 
 .label {
-  font-family: var(--mantine-font-family) !important;
   font-weight: 400;
   position: absolute;
   pointer-events: none;
