+++ src/ui/components/InputWithLabel/profile.module.scss
@@ -1,12 +1,10 @@
 .profile_input {
   background-color: #fcfcfd;
-  font-family: var(--mantine-font-family) !important;
 
   &::placeholder {
     color: var(--prune-text-gray-800);
     font-weight: 500;
     font-size: 14px !important;
-    font-family: var(--mantine-font-family) !important;
   }
 }
